+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Version 0.0.x
 
+### 0.0.7.x (2023-Oct-20)
+- Added implementation for state text for Multistate Input/Output/Value
+- Added implementation for getting Present Value and ObjectName for all missing objects
+
+### 0.0.6.x (2023-Oct-04)
+- Updated README.md with more detailed instructions for compiling and building the example. Made sure that all files required for compiling are mentioned. [Issue/6](https://github.com/chipkin/BACnetServerExamplePython/issues/6)
+- Tested using BACnet Version 4.1.19.2330
+
 ### 0.0.5.x (2022-Mar-07)
 - Updated error code on missing password for reinitialize device